test(deploy): run tool-execution scenarios against any recipe (OPS-8418)
The end-to-end tool-execution tests added in #13979 only ran against locally
spawned processes, because they lived inside the SGLang module and depended on
its fixtures. Nothing about the scenarios is local, though: they need only an
OpenAI-compatible client and a model name, and the tools they invoke run as
subprocesses in the pytest process rather than in the deployment. That is what
makes them portable.

Move the streaming helpers and both scenarios into tests/utils/tool_calling.py
and have the SGLang module import them, so the same assertions can be driven by
any topology. Behaviour there is unchanged: 26 tests collect as before, and the
xfail on the chained scenario keeps its reason.

Add tests/deploy/test_recipe_tool_execution.py, which deploys any
DynamoGraphDeployment manifest on any cluster and runs those scenarios against
it. Both are chosen at run time -- `--recipe <path>` and KUBECONFIG -- so no
recipe is special-cased.

Two gates keep the results honest:

  * Precondition: the manifest must configure a tool-call parser. Without one
    the frontend cannot emit tool_calls at all, so a failure would say nothing
    about the model. Missing parser skips before deploying, rather than burning
    a deployment to fail confusingly.
  * Capability: chaining two tools is a property of the model, not the serving
    stack, so it is recorded via record_property instead of asserted. Pointing
    this at a weaker recipe reports the limitation rather than manufacturing a
    red build.

The served model name comes from the manifest when it passes --model, and is
otherwise read back off /v1/models on the running deployment. Only 74 of the
148 recipe manifests name the model in args; the rest configure it through an
engine config or ConfigMap, where the endpoint is authoritative anyway.

Validation: single-node k3s-in-Docker with the NVIDIA device plugin, the
dynamo-platform chart, and an RTX 6000 Ada. Deploying a Qwen3-0.6B vLLM agg
DGD reached Ready in ~70s and both scenarios ran against it; two independent
runs agreed, with GPU memory peaking at 46 GiB, confirming the requests were
served by a GPU-backed worker in-cluster rather than a CPU fallback. The
single-tool scenario passed; the chained scenario reported NOT SUPPORTED,
passing 'alice' instead of the id the first tool returned -- the same
hallucinated-placeholder signature measured for Qwen3-0.6B locally. Running
the unmodified examples/backends/vllm/deploy/agg.yaml exercised the
precondition gate and skipped without deploying.
